Move reference genome implementation into separate module.
